remove AbstractResetIntegrationTest.java
